requirements.txt
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==2.0.0
adafruit-circuitpython-max31856==0.12.0
adafruit-circuitpython-requests==3.2.8
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.8.1
args==0.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
clint==0.5.1
Flask==2.3.2
Flask-Classful==0.14.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
paho-mqtt==1.6.1
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing-extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.16.1
